ELS indexer refactored a bit to allow for supporting queries that require a sequential series ELS searches.
