+++ server/lib/indexers/elasticsearch.js
@@ -1,10 +1,9 @@
 'use strict';
 
-const moment = require('moment');
 const elasticsearch = require('../elasticsearch');
 const {enforce} = require('../helpers');
 const interoperableErrors = require('../../../shared/interoperable-errors');
-const { SignalType, IndexMethod } = require('../../../shared/signals');
+const { IndexMethod } = require('../../../shared/signals');
 const { getIndexName, getFieldName, createIndex, extendMapping } = require('./elasticsearch-common');
 const contextHelpers = require('../context-helpers');
 
@@ -12,9 +11,10 @@ const signalSets = require('../../models/signal-sets');
 const em = require('../extension-manager');
 const fork = require('child_process').fork;
 
-const handlebars = require('handlebars');
 const path = require('path');
 const log = require('../log');
+
+const {query} = require('./elasticsearch-query');
 
 const insertBatchSize = 1000;
 
@@ -54,436 +54,6 @@ async function init() {
     for (const sigSet of sigSets) {
         await signalSets.index(contextHelpers.getAdminContext(), sigSet.id, IndexMethod.INCREMENTAL);
     }
-}
-
-// Converts an interval into elasticsearch interval
-function getElsInterval(duration) {
-    const units = ['ms', 's', 'm', 'h'];
-    for (const unit of units) {
-        if (duration.get(unit) !== 0) {
-            return duration.as(unit) + unit;
-        }
-    }
-
-    return duration.as('d') + 'd';
-}
-
-const aggHandlers = {
-    min: aggSpec => ({
-        id: 'min',
-        getAgg: field => ({
-            min: field
-        }),
-        processResponse: resp => resp.value
-    }),
-    avg: aggSpec => ({
-        id: 'avg',
-        getAgg: field => ({
-            avg: field
-        }),
-        processResponse: resp => resp.value
-    }),
-    max: aggSpec => ({
-        id: 'max',
-        getAgg: field => ({
-            max: field
-        }),
-        processResponse: resp => resp.value
-    }),
-    percentiles: aggSpec => ({
-        id: 'percentiles',
-        getAgg: field => ({
-            percentiles: {
-                percents: aggSpec.percents,
-                ...field
-            }
-        }),
-        processResponse: resp => resp.values
-    }),
-};
-
-function getAggHandler(aggSpec) {
-    let aggHandler;
-
-    if (typeof aggSpec === 'string') {
-        aggHandler = aggHandlers[aggSpec]
-
-        if (aggHandler) {
-            return aggHandler({});
-        } else {
-            throw new Error(`Invalid aggregation function ${aggSpec}`);
-        }
-    } else {
-        aggHandler = aggHandlers[aggSpec.type];
-        if (aggHandler) {
-            return aggHandler(aggSpec);
-        } else {
-            throw new Error(`Invalid aggregation function ${aggSpec.type}`);
-        }
-    }
-}
-
-function createElsQuery(query) {
-    const signalMap = query.signalMap;
-
-    function createElsScript(field) {
-        const fieldNamesMap = {};
-        for (const sigCid in signalMap) {
-            fieldNamesMap[sigCid] = getFieldName(signalMap[sigCid].id);
-        }
-
-        const scriptSource = field.settings.painlessScript;
-
-        // Handlebars replaces {{cid}} by the unique id of the signal
-        const scriptTemplate = handlebars.compile(scriptSource, {noEscape:true});
-        const scriptSubstituted = scriptTemplate(fieldNamesMap);
-        return {source: scriptSubstituted};
-    }
-
-    function getField(field) {
-        if (field.type === SignalType.PAINLESS) {
-            return { script: createElsScript(field) };
-        } else {
-            return { field: getFieldName(field.id) };
-        }
-    }
-
-    function createSignalAggs(signals) {
-        const aggs = {};
-
-        for (const sig in signals) {
-            for (const aggSpec of signals[sig]) {
-                const aggHandler = getAggHandler(aggSpec)
-                const sigFld = signalMap[sig];
-
-                if (!sigFld) {
-                    throw new Error(`Unknown signal ${sig}`);
-                }
-
-                const sigFldName = getFieldName(sigFld.id);
-
-                aggs[`${aggHandler.id}_${sigFldName}`] = aggHandler.getAgg(getField(sigFld));
-            }
-        }
-
-        return aggs;
-    }
-
-    function createElsAggs(aggs) {
-        const elsAggs = {};
-        let aggNo = 0;
-        for (const agg of aggs) {
-            const field = signalMap[agg.sigCid];
-            if (!field) {
-                throw new Error(`Unknown signal ${agg.sigCid}`);
-            }
-
-            const elsAgg = {
-            };
-
-            if (field.type === SignalType.DATE_TIME) {
-                // TODO: add processing of range buckets
-
-                elsAgg.date_histogram = {
-                    ...getField(field),
-                    interval: getElsInterval(moment.duration(agg.step)),
-                    offset: getElsInterval(moment.duration(agg.offset)),
-                    min_doc_count: agg.minDocCount
-                };
-
-            } else if (field.type === SignalType.INTEGER || field.type === SignalType.LONG || field.type === SignalType.FLOAT || field.type === SignalType.DOUBLE) {
-                elsAgg.histogram = {
-                    ...getField(field),
-                    interval: agg.step,
-                    offset: aggs.offset,
-                    min_doc_count: agg.minDocCount
-                };
-
-            } else {
-                throw new Error('Type of ' + agg.sigCid + ' (' + field.type + ') is not supported in aggregations');
-            }
-
-            if (agg.signals) {
-                elsAgg.aggs = createSignalAggs(agg.signals);
-            } else if (agg.aggs) {
-                elsAgg.aggs = createElsAggs(agg.aggs);
-            }
-
-            if (agg.order || agg.limit) {
-                elsAgg.aggs.sort = {
-                    bucket_sort: {}
-                };
-
-                if (agg.order) {
-                    elsAgg.aggs.sort.bucket_sort.sort = [
-                        {
-                            _key: {
-                                order: agg.order
-                            }
-                        }
-                    ];
-                }
-
-                if (agg.limit) {
-                    elsAgg.aggs.sort.bucket_sort.size = agg.limit;
-                }
-            }
-
-            elsAggs['agg_' + aggNo] = elsAgg;
-
-            aggNo += 1;
-        }
-
-        return elsAggs;
-    }
-
-    function createElsSort(sort) {
-        const elsSort = [];
-        for (const srt of sort) {
-            const field = signalMap[srt.sigCid];
-
-            if (!field) {
-                throw new Error('Unknown field ' + srt.sigCid);
-            }
-
-            elsSort.push({
-                [getFieldName(field.id)]: {
-                    order: srt.order
-                }
-            })
-        }
-
-        return elsSort;
-    }
-
-
-    const elsQry = {
-    };
-
-    const filter = [];
-    for (const range of query.ranges) {
-        const field = signalMap[range.sigCid];
-
-        if (!field) {
-            throw new Error('Unknown field ' + range.sigCid);
-        }
-
-        const rng = {};
-        const rngAttrs = ['gte', 'gt', 'lte', 'lt'];
-        for (const rngAttr of rngAttrs) {
-            if (range[rngAttr]) {
-                rng[rngAttr] = range[rngAttr];
-            }
-        }
-
-        filter.push({
-            range: {
-                [getFieldName(field.id)]: rng
-            }
-        });
-    }
-
-    elsQry.query = {
-        bool: {
-            filter
-        }
-    };
-
-
-    if (query.aggs) {
-        elsQry.size = 0;
-        elsQry.aggs = createElsAggs(query.aggs);
-
-    } else if (query.docs) {
-        if ('limit' in query.docs) {
-            elsQry.size = query.docs.limit;
-        }
-
-        elsQry._source = [];
-        elsQry.script_fields = {};
-        for (const sig of query.docs.signals) {
-            const sigFld = signalMap[sig];
-
-            if (!sigFld) {
-                throw new Error(`Unknown signal ${sig}`);
-            }
-
-            const sigFldName = getFieldName(sigFld.id);
-
-            const elsFld = getField(sigFld);
-            if (elsFld.field) {
-                elsQry._source.push(elsFld.field);
-            } else if (elsFld.script) {
-                elsQry.script_fields[sigFldName] = elsFld;
-            }
-        }
-
-        if (query.docs.sort) {
-            elsQry.sort = createElsSort(query.docs.sort);
-        }
-
-    } else if (query.sample) {
-        // TODO
-
-    } else if (query.summary) {
-        elsQry.size = 0;
-        elsQry.aggs = createSignalAggs(query.summary.signals);
-
-    } else {
-        throw new Error('None of "aggs", "docs", "sample", "summary" query part has been specified');
-    }
-
-    return elsQry;
-}
-
-function processElsQueryResult(query, elsResp) {
-    const signalMap = query.signalMap;
-
-    function processSignalAggs(signals, elsSignalsResp) {
-        const result = {};
-
-        for (const sig in signals) {
-            const sigBucket = {};
-            result[sig] = sigBucket;
-
-            const sigFldName = getFieldName(signalMap[sig].id);
-
-            for (const aggSpec of signals[sig]) {
-                const aggHandler = getAggHandler(aggSpec);
-                sigBucket[aggHandler.id] = aggHandler.processResponse(elsSignalsResp[`${aggHandler.id}_${sigFldName}`]);
-            }
-        }
-
-        return result;
-    }
-
-    function processElsAggs(aggs, elsAggsResp) {
-        const result = [];
-
-        let aggNo = 0;
-        for (const agg of aggs) {
-            const elsAggResp = elsAggsResp['agg_' + aggNo];
-
-            const buckets = [];
-
-            const field = signalMap[agg.sigCid];
-            if (field.type === SignalType.DATE_TIME) {
-                // TODO: add processing of range buckets
-
-                for (const elsBucket of elsAggResp.buckets) {
-                    buckets.push({
-                        key: moment.utc(elsBucket.key).toISOString(),
-                        count: elsBucket.doc_count
-                    });
-                }
-
-            } else if (field.type === SignalType.INTEGER || field.type === SignalType.LONG || field.type === SignalType.FLOAT || field.type === SignalType.DOUBLE) {
-                for (const elsBucket of elsAggResp.buckets) {
-                    buckets.push({
-                        key: elsBucket.key,
-                        count: elsBucket.doc_count
-                    });
-                }
-
-            } else {
-                throw new Error('Type of ' + agg.sigCid + ' (' + field.type + ') is not supported in aggregations');
-            }
-
-            if (agg.signals) {
-                let bucketIdx = 0;
-                for (const elsBucket of elsAggResp.buckets) {
-                    buckets[bucketIdx].values = processSignalAggs(agg.signals, elsBucket);
-                    bucketIdx += 1;
-                }
-
-            } else if (agg.aggs) {
-                let bucketIdx = 0;
-                for (const elsBucket of elsAggResp.buckets) {
-                    buckets[bucketIdx].aggs = processElsAggs(agg.aggs, elsBucket);
-                    bucketIdx += 1;
-                }
-            }
-
-            result.push(buckets);
-
-            aggNo += 1;
-        }
-
-        return result;
-    }
-
-    const result = {};
-
-    if (query.aggs) {
-        result.aggs = processElsAggs(query.aggs, elsResp.aggregations);
-
-    } else if (query.docs) {
-        result.docs = [];
-        result.total = elsResp.hits.total;
-
-        for (const hit of elsResp.hits.hits) {
-            const doc = {};
-
-            for (const sig of query.docs.signals) {
-                const sigFld = signalMap[sig];
-
-                if (sigFld.type === SignalType.PAINLESS) {
-                    doc[sig] = hit.fields[getFieldName(sigFld.id)][0];
-                } else {
-                    doc[sig] = hit._source[getFieldName(sigFld.id)];
-                }
-            }
-
-            result.docs.push(doc);
-        }
-
-    } else if (query.sample) {
-        // TODO
-
-    } else if (query.summary) {
-        result.summary = processSignalAggs(query.summary.signals, elsResp.aggregations);
-
-    } else {
-        throw new Error('None of "aggs", "docs", "sample", "summary" query part has been specified');
-    }
-
-    return result;
-}
-
-
-async function query(queries) {
-    const msearchBody = [];
-
-    for (const sigSetQry of queries) {
-        const indexName = getIndexName(sigSetQry.sigSet);
-
-        msearchBody.push({index: indexName});
-        msearchBody.push(createElsQuery(sigSetQry));
-    }
-
-    const msearchResult = await elasticsearch.msearch({body:msearchBody});
-
-    // FIXME - process errors caused by asking for too many data points - this should throw interoperableErrors.TooManyPointsError
-
-    const errorResponses = [];
-    for (const response of msearchResult.responses) {
-        if (response.error) {
-            errorResponses.push(JSON.stringify(response.error));
-        }
-    }
-    if (errorResponses.length > 0) {
-        log.error("Indexer", "Elasticsearch queries failed");
-        log.verbose(errorResponses);
-        throw new Error("Elasticsearch queries failed");
-    }
-
-    const result = [];
-
-    for (let idx = 0; idx < queries.length; idx++) {
-        result.push(processElsQueryResult(queries[idx], msearchResult.responses[idx]));
-    }
-
-    return result;
 }
